refactor: restructure src into focused modules
Split the 1400-line prompts.ts monolith into prompts/, extract domain
logic into permissions/resolve and auth/template-url, move CLI flags and
interactive create flow out of index.ts, and keep prompts.ts as a shim.

Closes #145, #146, #147
