+++ src/flows/token-creation-flow-error.ts
@@ -1,0 +1,3 @@
+import { TokenCreationFlowErrorBase } from "#src/tagged-error-bases.ts";
+
+export class TokenCreationFlowError extends TokenCreationFlowErrorBase {}
