+++ src/flows/token-deletion-flow-error.ts
@@ -1,0 +1,3 @@
+import { TokenDeletionFlowErrorBase } from "#src/tagged-error-bases.ts";
+
+export class TokenDeletionFlowError extends TokenDeletionFlowErrorBase {}
