+++ src/prompts/flow/post-create.ts
@@ -1,0 +1,36 @@
+import { note, select } from "@clack/prompts";
+
+import { CF_API_TOKENS_URL } from "#src/auth/template-url.ts";
+import colour from "#src/colour.ts";
+import { check, exitIfNonInteractive } from "#src/prompts/guards.ts";
+import type { PostCreateAction } from "#src/prompts/types.ts";
+
+export async function askPostCreateAction(): Promise<PostCreateAction> {
+  exitIfNonInteractive();
+  return check(
+    await select({
+      initialValue: "done",
+      message: "What should we do with the key you just created?",
+      options: [
+        { label: "Keep and finish", value: "done" },
+        { label: "Replace this key", value: "revoke-again" },
+        { label: "Delete this key", value: "revoke-done" },
+        { label: "Keep and create another", value: "again" },
+      ],
+    })
+  ) as PostCreateAction;
+}
+
+/**
+ * Display the newly created token value in a note box.
+ * The value is only returned by the API on creation and cannot be retrieved again.
+ *
+ * @param tokenValue - The raw token secret returned by the API.
+ * @param tokenName - The display name of the created token.
+ */
+export function showCreatedToken(tokenValue: string, tokenName: string): void {
+  note(
+    `${colour.CYAN}${tokenValue}${colour.RESET}\n\n${colour.WHITE}Copy this token now.${colour.RESET} It will not be shown again.\nManage tokens: ${colour.CYAN}${CF_API_TOKENS_URL}${colour.RESET}`,
+    `Token created: ${tokenName}`
+  );
+}
